refactor(agent): extract agent + rpc modules and finish the split
Final split. linux/mod.rs is now a 16-line index of submodules with
a single `pub use agent::Agent` re-export.

- agent.rs: Agent struct, impl, Default. Owns Agent::run — the vsock
  accept loop, the eager sandbox-service init, and spawning the two
  API proxies. Calls into rpc::handle_connection per accepted stream.
- rpc.rs: per-connection dispatch (handle_connection), the inline
  RpcRequest router (handle_request), the small handlers (Ping +
  Shutdown + MmapReadFile), and sync_clock_from_host (still
  pub(super) because runtime::ensure_runtime_prerequisites needs it).
  RequestResult lives here too, next to its sole producer / consumer.

Cross-module wiring: agent.rs / stub.rs / runtime.rs now import the
items they actually need by absolute path (`arcbox_constants::ports::
AGENT_PORT`, `crate::agent::ensure_runtime`, `super::rpc::sync_clock_
from_host`) instead of going through linux/mod.rs as a re-export hub.

After this commit the Linux implementation is laid out as:

  agent/linux/
    mod.rs              — 11 mod declarations + pub use Agent
    agent.rs            — Agent struct, run loop, proxy spawn
    rpc.rs              — connection dispatch + ping/shutdown/mmap
    runtime.rs          — containerd + dockerd lifecycle
    kubernetes.rs       — k3s lifecycle + 5 K8s RPCs
    sandbox.rs          — SandboxService dispatch
    system_info.rs      — /proc collector
    proxy.rs            — Docker / K8s API vsock proxies
    btrfs.rs            — data volume detection / format / mount
    cmdline.rs          — /proc/cmdline + env readers
    probe.rs            — TCP / Unix / Docker API probes
    vsock.rs            — listener bind + peer-closed classifier

Original agent/mod.rs (2420 lines, 14 inline section dividers)
ends as a 132-line shell that just declares `mod linux` / `mod stub`,
re-exports `Agent`, and hosts the docker-log-line tests.

cargo check (host stub + aarch64-unknown-linux-musl) and cargo clippy
both produce the same warning set as master (5 pre-existing dead-code /
unused-import warnings; no new findings introduced by the refactor).
